refactor: remove index file from private module
