--- src/app/wallets/private/index.ts
@@ -1,2 +1,0 @@
-export * from "./check-limit-helpers"
-export * from "./get-on-chain-txs"
