+++ src/app/wallets/send-on-chain.ts
@@ -40,7 +40,7 @@ import {
   checkAndVerifyTwoFA,
   checkIntraledgerLimits,
   checkWithdrawalLimits,
-} from "./private"
+} from "./private/check-limit-helpers"
 import { getOnChainFee } from "./get-on-chain-fee"
 
 const { dustThreshold } = getOnChainWalletConfig()
